rewrited Board
